amlhalasink: desotry ESSOS RM during ready to null transition [1/1]
The audio sink requests the resource during NULL to READY,
but releases it during PAUSED to READY.  It seems there must
be some runtime scenarios where the audio sink is cycled
between NULL to READY and READY to NULL a couple of times
which leads calling EssRMgrCreate a second time without an
intervening call to EssRMgrDestroy or EssRMgrReleaseResource.

If an audio sink element is taken from NULL to READY and then
READY to NULL and then disposed of, then the next time a new audio
sink instance requests the audio decoder resource it will cause the
“current owner” to be called with a revoke notification but the
user data will be bad – pointing to a destroyed audio sink instance
with resulting memory corruption and/or crash.

Change-Id: I240a87d024a1119403e3018bb6397d9d3191f7a9