Yocto: AVTS 3.0 ShouldPlayUntilFinishedAfterVideoBufferU... [1/1]
PD#SWPL-65356

Problem:
When the player status switch to pause and audio source element data
just finishes playing, audio source element will issue the
stream_group_done event in the paused state. After receiving the event,
asink will hal_stop and destroy avsync, resulting in a small amount of
audio data cached in MS12 is not played. so the test item finally
failed because the playing time was not long enough

Solution:
When the player state is paused, the pcr cannot be obtained through
the avsync_get_time interface, so it is considered that the player
has not started playing, and it returns directly. Add non-pause state
conditions

Verify:
yocto-ap222

Change-Id: I87cdeee33a7ddc4d197f0812c108c8ce6fd82a5b
Signed-off-by: shu.zhang <shu.zhang@amlogic.com>