Yocto: MPEG1 video playback cannot exit [1/2]
PD#OTT-27010

Problem:
When the length of audio in a multimedia file is far less than
that of video, asink will not send data to Audio Hal after the
audio is played, and xrun will occur. When xrun occurs, avsync
will be paused, so that the video cannot be played normally.

Solution:
When a Gap Event is received, Xrun is paused

Verify:
yocto-ap222

Change-Id: Ib18e62b4fe4049a7bd90f186308c57b60f6a0bbe
Signed-off-by: shu.zhang <shu.zhang@amlogic.com>